feat: ensure url provided contains protocol
